requirements.txt
requests
RPi.GPIO
picamera
pickle5
opencv-python
paho-mqtt
pybluez
scipy